More functionality in the roadbuild tools
Straight for road and rail:
Based on the start and end MapProject, a dot will appear that indicates a
straight connection from the selected MapProject.

Curved connection ProjectState:
When clicking on any MapProject and Ground or Intersection, a straight
road would be created. Now an interpolation point needs to be set.

Function for line line intersection
With four points, representing two lines, the intersection of these two
lines is calculated

Function for line closed point
Given two points representing one line and one point, the closed point
on the line from the third point is calculated.